pbrd: nexthop refcounting
Keep a hash of nexthops used by the NHT subsystem of pbrd together with
a refcount for each nexthop. Only send a removal message to Zebra when
the refcount drops to zero.

As part of this, change the nexthop member of pbr_nexthop_cache to a
pointer to the single copy stored in the hash.

Signed-off-by: Quentin Young <qlyoung@cumulusnetworks.com>