vrrpd: continue to flesh out Lua API
Instead of relying on predefined function names, we are going to treat
each Lua script file as a single function. This is the natural thing to
do because the Lua C API compiles a chunk into a Lua function pointer.
We associate this function pointer with VRRP (object, instance) 2-tuple
and call it when some object tracking event occurs, providing it with an
environment of relevant information stored in objects that can be
manipulated to influence the underlying protocol daemon.

Best of all you don't even need to restart FRR to load changes to the
Lua scripts, nor will an error in the script crash FRR. Just edit the
script, write it to disk and FRR will use it.

Signed-off-by: Quentin Young <qlyoung@cumulusnetworks.com>