lib: change object key to just its name
The idea of (type, name) was nice but this ultimately means users will
have to enter both a freeform type and freeform name on the CLI, which
sucks from a usability perspective; you shouldn't have to know the types
of objects in advance, just their names. Consequently tracked objects
are now identified by name only, and we will warn if the type of an
object changes.

This also allows registering for object tracking with just a name before
the objects are being tracked in the first place, e.g. tracking
interfaces by name before we have connected to Zebra and have any kind
of authoritative state information on them.

Signed-off-by: Quentin Young <qlyoung@cumulusnetworks.com>